// apps/web/__generated__/userListSearchFragment.graphql.ts
export type userListSearchFragment$data = {
  readonly users: {
    readonly edges: ReadonlyArray<{
      readonly cursor: string;
      readonly node: {
        readonly id: string;
        readonly " $fragmentSpreads": FragmentRefs<"userItemFragment">;
      } | null | undefined;
    } | null | undefined>;
  };
  readonly " $fragmentType": "userListSearchFragment";
};
export type userListSearchFragment$key = {
  readonly " $data"?: userListSearchFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"userListSearchFragment">;
};